+++ src/lib/services/id.service.jsx
@@ -1,0 +1,5 @@
+// ID Service
+
+export const convertId = (id) => {
+    return (id === null || id ==='' || typeof id === 'string' ? id : id.$oid);
+}